Moved Implementations to functions.cpp
Moved all implementations to functions.cpp and kept function declarations in functions.h file.
